# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Speech Emotion Recognition as a Service
# during dev, should point to your local machine
host: localhost:5000
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - errorCode
      - message
    properties:
      errorCode:
        type: number
      message:
        type: string
  EmotionStatisticObject:
    required:
      - emotion
      - probability
    properties:
      emotion:
        type: string
      probability:
        type: number
  EmotionalStatisticsResponse:
    required:
      - emotions
    properties:
      emotions:
        type: array
        items:
          $ref: "#/definitions/EmotionStatisticObject"
paths:
  /swagger:
    x-swagger-pipe: swagger_raw
  /analyse/{userId}/{emotions}:
    x-swagger-router-controller: analyse
    post:
      consumes:
        - multipart/form-data
      description: |
        Send an audio file with speech and 
        get the emotional statistics shown 
        back (done via Support Vector 
        Machine Classification)
      operationId: analyseAll
      parameters:
        - name: userId
          in: path
          description: The SERaaS user account ID to query the endpoint with
          required: true
          type: string
        - name: emotions
          in: path
          description: |
            Emotions to be analysed, can be "all" 
            or a specific emotions. Multiple emotions 
            can be queried by comma i.e. "happy,sad" 
            Full list of emotions defined in Web 
            Platform
          required: true
          type: string
          default: all
        - name: file
          in: formData
          description: The audio file with speech to upload
          required: true
          type: file
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EmotionalStatisticsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /analyse/{userId}/{emotions}/{period}:
    x-swagger-router-controller: analyse
    post:
      consumes:
        - multipart/form-data
      description: |
        Send an audio file with speech and 
        get the emotional statistics shown 
        back (done via Support Vector 
        Machine Classification), analysing 
        the audio file periodically according 
        to the given timeframe.
      operationId: analyseAll
      parameters:
        - name: userId
          in: path
          description: The SERaaS user account ID to query the endpoint with
          required: true
          type: string
        - name: emotions
          in: path
          description: |
            Emotions to be analysed, can be "all" 
            or a specific emotions. Multiple emotions 
            can be queried by comma i.e. "happy,sad" 
            Full list of emotions defined in Web 
            Platform
          required: true
          type: string
          default: all
        - name: period
          in: path
          description: |
            The periodic length of the audio file to 
            be analysed. E.g. if 2 is specified, the 
            audio file's emotions are analysed in 2 
            second intervals.
          required: true
          type: number
          default: -1
        - name: file
          in: formData
          description: The audio file with speech to upload
          required: true
          type: file
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EmotionalStatisticsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"